backend/requirements.txt
Django==5.1.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.2
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
python-decouple==3.8
cryptography==41.0.7
Pillow==10.1.0
opencv-python==4.8.1.78
ultralytics==8.0.224
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
requests==2.31.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
drf-spectacular==0.27.0
django-filter==23.4
django-extensions==3.2.3
python-dateutil==2.8.2
pydantic==2.5.0
